refactor(optimizer): split up src/lib
