+++ packages/optimizer/src/commons.d.ts
