+++ packages/optimizer/src/commons.ts
@@ -1,0 +1,15 @@
+import type { RuleNode, RawPublicodes } from 'publicodes'
+
+export type RuleName = string
+export type ParsedRules = Record<RuleName, RuleNode<RuleName>>
+export type RawRules = RawPublicodes<RuleName>
+
+export function getRawNodes(parsedRules: ParsedRules): RawRules {
+	return Object.fromEntries(
+		Object.values(parsedRules).reduce((acc, rule) => {
+			const { nom, ...rawNode } = rule.rawNode
+			acc.push([nom, rawNode])
+			return acc
+		}, [])
+	)
+}
