+++ packages/optimizer/test/utils.test.ts
@@ -1,7 +1,10 @@
-import type { ParsedRules, RawRules } from '../src/lib'
+import type { ParsedRules, RawRules } from '../src/commons'
 import Engine from 'publicodes'
 
-export function callWithEngine<R>(fn: (engine) => R, rawRules: RawRules): R {
+export function callWithEngine<R>(
+	fn: (engine: Engine) => R,
+	rawRules: RawRules
+): R {
 	const engine = new Engine(rawRules)
 	return fn(engine)
 }
